refactor; cleanup; bump to 0.2.9 (untested)
